# src/main/resources/api.yml
openapi: 3.1.0
info:
  version: v1
  title: content-vault API
  description: content-vault RESTFul api for openapi 3.1
  termsOfService: 'https://jumkid.com/terms-of-use'
  contact:
    name: Chooli Yip
    url: jumkid.com
    email: chooli.yip@gmail.com
  license:
    name: Jumkid Public License
    url: 'http://license.jumkid.com'
servers:
  - description: Testing
    url: 'http://test.jumkid.com'
  - description: Live production
    url: 'http://api.jumkid.com'
...
security:
  - bearerAuth: []
paths:
  /metadata:
    get:
      tags:
        - v1
      operationId: searchMetadata
      description: search metadata with query and size
      parameters:
        - name: q
          in: query
          required: false
          description: query string
          schema:
            type: string
        - name: size
          in: query
          required: false
          description: size of result set
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaFile'
    post:
      tags:
        - v1
      operationId: addMetadata
      description: save a new metadata
      parameters:
        - name: mediaFileModule
          in: query
          required: true
          description: mediaFile module
          schema:
            $ref: '#/components/schemas/MediaFileModule'    
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaFile'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFile'
      summary: Add raw metadata
  '/metadata/{id}':
    get:
      tags:
        - v1
      operationId: getMetadata
      description: get single metadata for media file
      parameters:
        - name: id
          in: path
          required: true
          description: mediafile identifier
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFile'
    put:
      tags:
        - v1
      operationId: updateMetadata
      description: update metadata for media file
      parameters:
        - name: id
          in: path
          required: true
          description: mediafile identifier
          schema:
            type: string
      requestBody:
        description: partial metadata json object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaFile'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFile'

    delete:
      tags:
        - v1
      operationId: deleteMetadata
      description: delete single metadata for media file
      parameters:
        - name: id
          in: path
          required: true
          description: mediafile identifier
          schema:
            type: string
            
      responses:
        '204':
          description: successfully delete the metadata by givn id
          content:
            application/json:
              schema:
                type: integer
...
  '/gallery':
    post:
      tags:
        - v1
      operationId: saveGallery
      description: save multiple file as gallery
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
        - name: content
          in: query
          required: false
          schema:
            type: string
        - name: files
          in: query
          required: false
          schema:
            type: array
            format: file
        - name: tags
          in: query
          required: false
          schema:
            type: array
            format: string
        - name: accessScope
          in: query
          required: true
          schema:
            $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/AccessScope    
            
      responses:
        '202':
          description: save a gallery with multiple files upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFile'
  '/gallery/{id}':
    post:
      tags:
        - v1
      operationId: updateItems
      description: update a gallery with file upload
      parameters:
        - name: id
          in: path
          required: true
          description: gallery identifier
          schema:
            type: string
        - name: featuredId
          in: query
          required: false
          description: representative image of the gallery
          schema:
            type: string
        - name: files
          in: query
          required: false
          schema:
            type: array
            format: file

      responses:
        '202':
          description: updated a gallery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFile'

    put:
      tags:
        - v1
      operationId: update
      description: update a gallery with partialMediaFile or child media file list
      parameters:
        - name: id
          in: path
          required: true
          description: gallery identifier
          schema:
            type: string

      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaFile'

      responses:
        '202':
          description: updated a gallery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFile'

    delete:
      tags:
        - v1
      operationId: delete
      description: delete items from the gallary or the entire gallery
      parameters:
        - name: id
          in: path
          required: true
          description: gallery identifier
          schema:
            type: string
        - name: items
          in: query
          required: false
          schema:
            type: array
            format: string    

      responses:
        '202':
          description: successfully delete items from the gallery
          content:
            application/json:
              schema:
                type: array
                format: '#/components/schemas/MediaFile'
  '/gallery/{id}/clone':
    post:
      tags:
        - v1
      operationId: clone existing gallery
      description: clone a gallery with all properties and children
      parameters:
        - name: id
          in: path
          required: true
          description: gallery identifier
          schema:
            type: string
        - name: title
          in: query
          required: false
          description: title of gallery
          schema:
            type: string
        - name: toMediaGalleryId
          in: query
          required: false
          description: target empty media gallery id
          schema:
            type: string    

      responses:
        '202':
          description: accpeted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFile'
...
components:
  schemas:
    MediaFile:
      description: Model of media file
      properties:
        id:
          type: string
          example: 1mWG-nQBWmmN0po1BUoG
        filename:
          type: string
          example: foo.txt
        mimeType:
          type: string
          example: text/plain
        size:
          type: integer
          description: file size in byte
        title:
          type: string
        content:
          type: string
        activated:
          type: boolean
        accessScope:
          $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/AccessScope
        createdBy:
          type: string
        creationDate:
          $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime
        modifiedBy:
          type: string
        modifiedOn:
          $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime
        props:
          type: array
          items:
            description: list of additional properties
            properties:
              name:
                type: string
              textValue:
                type: string
              dateValue:
                $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime
              numberValue:
                type: integer
        tags:
          type: array
          items:
            description: list of tags
            type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/MediaFile'
...
    MediaFileModule:
      type: string
      enum:
        - FILE
        - TEXT
        - HTML
        - GALLERY
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT